Move pool_snapshot
