+++ include/edyn/networking/context/server_network_context.hpp
@@ -4,7 +4,7 @@
 #include <vector>
 #include <entt/entity/fwd.hpp>
 #include <entt/signal/sigh.hpp>
-#include "edyn/networking/packet/util/pool_snapshot.hpp"
+#include "edyn/networking/util/pool_snapshot.hpp"
 #include "edyn/networking/util/server_pool_snapshot_importer.hpp"
 #include "edyn/networking/util/server_pool_snapshot_exporter.hpp"
