+++ include/edyn/networking/packet/general_snapshot.hpp
@@ -2,7 +2,7 @@
 #define EDYN_NETWORKING_PACKET_GENERAL_SNAPSHOT_HPP
 
 #include <vector>
-#include "edyn/networking/packet/util/pool_snapshot.hpp"
+#include "edyn/networking/util/pool_snapshot.hpp"
 
 namespace edyn::packet {
