+++ include/edyn/networking/packet/transient_snapshot.hpp
@@ -2,7 +2,7 @@
 #define EDYN_NETWORKING_PACKET_TRANSIENT_SNAPSHOT_HPP
 
 #include <vector>
-#include "edyn/networking/packet/util/pool_snapshot.hpp"
+#include "edyn/networking/util/pool_snapshot.hpp"
 
 namespace edyn::packet {
