+++ include/edyn/networking/util/pool_snapshot.hpp
